完成提交考试答案接口,添加https支持

+++ b2c-parent/b2c-core/src/main/java/org/javamaster/b2c/core/entity/ExamsRecord.java
@@ -30,7 +30,7 @@ public class ExamsRecord {
     /**
      * 得分
      */
-    private Byte score;
+    private Short score;
 
     /**
      * 创建时间
@@ -97,14 +97,14 @@ public class ExamsRecord {
     /**
      * 获取得分
      */
-    public Byte getScore() {
+    public Short getScore() {
         return score;
     }
 
     /**
      * 设置得分
      */
-    public void setScore(Byte score) {
+    public void setScore(Short score) {
         this.score = score;
     }
